added all game logic and result page
