+++ index.html
@@ -7,7 +7,7 @@
     </head>
     <body>
         <div class="difficulties center">
-         
+           
         </div>
     </body>
     <script type="module" src="./level-pages/index.js"></script>
